+++ app/src/main/java/org/simple/clinic/storage/migrations/RoomMigrationsModule.kt
@@ -107,7 +107,8 @@ class RoomMigrationsModule {
       migration99: Migration_99,
       migration100: Migration_100,
       migration101: Migration_101,
-      migration102: Migration_102
+      migration102: Migration_102,
+      migration103: Migration_103
   ): List<Migration> {
     return listOf(
         migration_3_4,
@@ -208,7 +209,8 @@ class RoomMigrationsModule {
         migration99,
         migration100,
         migration101,
-        migration102
+        migration102,
+        migration103
     )
   }
 }