+++ CHANGELOG.md
@@ -5,8 +5,6 @@
 ### Internal
 
 - Fix inconsistent overdue section header UI
-- Use `OverdueListSectionStates` to update overdue section states in `OverdueModel`
-- Use `OverdueListSectionStates` to check overdue section states in `OverdueAppointmentListItemNew`
 - Bump sentry-android to v6
 - Bump sentry android gradle plugin to v3.1.0
 - Bump Google Play Services Location to v20
@@ -14,6 +12,10 @@
 - Bump Sentry to v6.1.0
 - Bump Kotlin to v1.7.0
 - Hide footer when the pending section is collapsed
+- Restore expanded/collapsed states of Overdue Sections
+  - Use `OverdueListSectionStates` to update overdue section states in `OverdueModel`
+  - Use `OverdueListSectionStates` to check overdue section states in `OverdueAppointmentListItemNew`
+  - Remove unused section header state usages
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-13-8291