+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - Use `OverdueListSectionStates` to check overdue section states in `OverdueAppointmentListItemNew`
   - Remove unused section header state usages
 - Add developer flag for overdue search
+- Add search button in overdue patients list
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-13-8291