+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Add developer flag for overdue search
 - Add search button in overdue patients list
 - Implement FTS table for `PatientAddress`
+- Define overdue appointment query as a `const` in `OverdueAppointment`
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-13-8291