+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Add search button in overdue patients list
 - Implement FTS table for `PatientAddress`
 - Define overdue appointment query as a `const` in `OverdueAppointment`
+- Bump Annotation library to v1.4.0
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-13-8291